backend/requirements.txt
fastapi
uvicorn
pydantic
openai
supabase
python-dotenv
python-multipart
requests
